+++ app/Http/Controllers/NewsItemController.php
@@ -2,7 +2,7 @@
 
 namespace App\Http\Controllers;
 
-use App\NewsItem;
+use App\Models\NewsItem;
 use App\Repositories\NewsItemRepository;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Config;
