+++ database/factories/NewsItemFactory.php
@@ -2,7 +2,7 @@
 
 namespace Database\Factories;
 
-use App\NewsItem;
+use App\Models\NewsItem;
 use Faker\Generator as Faker;
 
 $factory->define(NewsItem::class, function (Faker $faker) {
