+++ database/seeders/ClearSeeder.php
@@ -10,8 +10,8 @@ use App\Models\ApplicationForm\ApplicationFormRow;
 use App\Models\Book;
 use App\Models\Certificate;
 use App\Models\MenuItem;
+use App\Models\NewsItem;
 use App\Models\UserEventLogEntry;
-use App\NewsItem;
 use App\Rol;
 use App\User;
 use App\Zekering;
