+++ tests/Feature/News/CreateNewsItemTest.php
@@ -2,7 +2,7 @@
 
 namespace Tests\Feature\News;
 
-use App\NewsItem;
+use App\Models\NewsItem;
 use App\User;
 use Illuminate\Foundation\Testing\DatabaseMigrations;
 use Illuminate\Support\Facades\Artisan;
